eigenstate - landgreen catch-up

+++ js/entities/bullet.js
@@ -491,7 +491,7 @@ const b = {
             knock = Vector.mult(Vector.normalise(sub), -Math.sqrt(dmg) * body[i].mass * 0.022);
             body[i].force.x += knock.x;
             body[i].force.y += knock.y;
-            if (tech.isBlockExplode) {
+            if (tech.isBlockExplode && !body[i].isInvulnerable) {
               if (body[i] === m.holdingTarget) m.drop()
               const size = 20 + 300 * Math.pow(body[i].mass, 0.25)
               const where = body[i].position
@@ -515,7 +515,7 @@ const b = {
         if (bullet[i].bulletType === "foam") {
           try {
             sub = Vector.sub(where, bullet[i].position);
-            dist = Vector.magnitude(sub);
+            dist = Vector.magnitude(sub) - bullet[i].radius;
             if (dist < radius && tech.isFoamExplode && !bullet[i].hasExploded) { //&& Math.random() < 0.63
               const bubblePos = bullet[i].position;
               const size = 20 + 150 * Math.pow(bullet[i].radius, 0.25);
@@ -1577,7 +1577,7 @@ const b = {
             for (let i = 0, len = powerUp.length; i < len; ++i) {
               if (powerUp[i] === this.caughtPowerUp) index = i
             }
-            if (index !== null) {
+            if (index !== null && !simulation.paused) {
               powerUps.onPickUp(this.caughtPowerUp);
               this.caughtPowerUp.effect();
               Matter.Composite.remove(engine.world, this.caughtPowerUp);
@@ -1693,7 +1693,7 @@ const b = {
               for (let i = 0; i < blocks.length; i++) {
                 if (blocks[i].bodyA.classType === "body" && !blocks[i].bodyA.isNotHoldable && blocks[0].bodyA.mass < 40) {
                   this.retract()
-                  if (tech.hookNails) {
+                  if (tech.hookNails && !body[i].isInvulnerable) {
                     b.targetedNail(this.position, 3 * tech.hookNails)
                     const ANGLE = 2 * Math.PI * Math.random() //make a few random ones
                     for (let i = 0; i < 13; i++) b.nail(this.position, { x: 10.5 * Math.cos(ANGLE), y: 10.5 * Math.sin(ANGLE) }, 1.2)
@@ -1997,7 +1997,7 @@ const b = {
           for (let i = 0, len = powerUp.length; i < len; ++i) {
             if (powerUp[i] === this.caughtPowerUp) index = i
           }
-          if (index !== null) {
+          if (index !== null && !simulation.paused) {
             powerUps.onPickUp(this.caughtPowerUp);
             this.caughtPowerUp.effect();
             Matter.Composite.remove(engine.world, this.caughtPowerUp);
@@ -2597,7 +2597,7 @@ const b = {
           bestPowerUp = vertexCollision(path[path.length - 2], path[path.length - 1], [mob, map, body, powerUp]);
           if (bestPowerUp.who) {
               for (let i = 0, len = powerUp.length; i < len; ++i) {
-                  if (powerUp[i] === bestPowerUp.who && !simulation.isChoosing && (powerUp[i].name !== "heal" || m.maxHealth - m.health > 0.01 || tech.isOverHeal) && !(tech.isEnergyNoAmmo && powerUp[i].name === "ammo")) {
+                  if (powerUp[i] === bestPowerUp.who && !simulation.isChoosing && !simulation.paused && (powerUp[i].name !== "heal" || m.maxHealth - m.health > 0.01 || tech.isOverHeal) && !(tech.isEnergyNoAmmo && powerUp[i].name === "ammo")) {
                       m.energy += 0.8
                       powerUps.onPickUp(powerUp[i]);
                       powerUp[i].effect();
@@ -3480,7 +3480,7 @@ const b = {
               const distB = Vector.magnitude(Vector.sub(this.position, b.position))
               return distA < distB ? a : b
             })
-            if (found && m.energy > 0.041) {
+            if (found && m.energy > 0.041 && !found.isInvulnerable) {
               m.energy -= 0.04
               m.fieldUpgrades[4].endoThermic(0.4)
               //remove the body and spawn a new drone
@@ -3588,7 +3588,7 @@ const b = {
                 const distB = Vector.magnitude(Vector.sub(this.position, b.position))
                 return distA < distB ? a : b
               })
-              if (found) this.bodyTarget = found
+              if (found && !found.isInvulnerable) this.bodyTarget = found
             }
           } else {
             this.do = this.doDieing
@@ -3742,7 +3742,7 @@ const b = {
               const distB = Vector.magnitude(Vector.sub(this.position, b.position))
               return distA < distB ? a : b
             })
-            if (found && m.energy > 0.091) {
+            if (found && m.energy > 0.091 && !found.isInvulnerable) {
               m.energy -= 0.09
               m.fieldUpgrades[4].endoThermic(0.7)
               //remove the body and spawn a new drone
@@ -3859,7 +3859,7 @@ const b = {
                       (m.health > 0.93 * m.maxHealth && !tech.isDroneGrab && powerUp[i].name === "heal") ||
                       (tech.isSuperDeterminism && powerUp[i].name === "field") ||
                       ((tech.isEnergyNoAmmo || b.inventory.length === 0) && powerUp[i].name === "ammo")
-                    )
+                    ) && !simulation.paused
                   ) {
                     //draw pickup for a single cycle
                     ctx.beginPath();
@@ -3893,7 +3893,7 @@ const b = {
                     (tech.isSuperDeterminism && powerUp[i].name === "field") ||
                     ((tech.isEnergyNoAmmo || b.inventory.length === 0) && powerUp[i].name === "ammo")
                   )) {
-                    if (Vector.magnitudeSquared(Vector.sub(this.position, powerUp[i].position)) < 20000 && !simulation.isChoosing) {
+                    if (Vector.magnitudeSquared(Vector.sub(this.position, powerUp[i].position)) < 20000 && !simulation.isChoosing && !simulation.paused) {
                       //draw pickup for a single cycle
                       ctx.beginPath();
                       ctx.moveTo(this.position.x, this.position.y);
@@ -5589,7 +5589,7 @@ const b = {
           for (let i = 0, len = powerUp.length; i < len; ++i) {
             if (powerUp[i] === this.caughtPowerUp) index = i
           }
-          if (index !== null) {
+          if (index !== null && !simulation.paused) {
             powerUps.onPickUp(this.caughtPowerUp);
             this.caughtPowerUp.effect();
             Matter.Composite.remove(engine.world, this.caughtPowerUp);
